# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Guest API",
        "description": "API Endpoints for managing guests",
        "version": "1.0.0"
    },
    "paths": {
        "/api/guests": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Retrieve a list of guests",
                "operationId": "56cc8fe33a100159d774722901ae37e8",
                "responses": {
                    "200": {
                        "description": "List of guests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guest"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Create a new guest",
                "operationId": "e792b91966348e6bb3439a3d24883e1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john.doe@example.com"
                                        },
                                        "phone_number": {
                                            "type": "string",
                                            "example": "+1234567890"
                                        },
                                        "country_code": {
                                            "type": "string",
                                            "example": "US",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "phone_number": [
                                                    "The phone number must be valid."
                                                ],
                                                "country_code": [
                                                    "The country code does not match the phone number."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guests/{id}": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Retrieve a specific guest by ID",
                "operationId": "4267610a6a43088364913e1b6071b25d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Update an existing guest",
                "operationId": "af858a767bd4c0c6e7fd978a27775871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john.doe@example.com"
                                        },
                                        "phone_number": {
                                            "type": "string",
                                            "example": "+1234567890"
                                        },
                                        "country_code": {
                                            "type": "string",
                                            "example": "US",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "phone_number": [
                                                    "The phone number must be valid."
                                                ],
                                                "email": [
                                                    "The email has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Delete a guest",
                "operationId": "627107a5c10ad7a3877e6ca545859ca8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guest deleted successfully"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GuestRequest": {
                "title": "Guest Request",
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "phone_number"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "phone_number": {
                        "description": "Must be a valid phone number.",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "country_code": {
                        "description": "Must match the phone number's country.",
                        "type": "string",
                        "maxLength": 2,
                        "example": "US",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Guest": {
                "title": "Guest",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "country_code": {
                        "type": "string",
                        "example": "US",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Guests",
            "description": "Guests"
        }
    ],
    "security": [
        {
            "Bearer": []
        }
    ]
}